requirements.txt
aiopg==1.3.4
anyio==3.6.1
async-timeout==4.0.2
asyncpg==0.26.0
click==8.1.3
databases==0.6.1
fastapi==0.81.0
greenlet==1.1.3
h11==0.13.0
idna==3.3
psycopg2-binary==2.9.3
pydantic==1.10.1
python-dotenv==0.20.0
sniffio==1.2.0
SQLAlchemy==1.4.40
starlette==0.19.1
typing_extensions==4.3.0
uvicorn==0.18.3
